alx_travel_app/requirements.txt
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.6.15
chapa==0.1.2
charset-normalizer==3.4.2
django==4.2.23
djangorestframework==3.15.2
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
typing-extensions==4.13.2
urllib3==2.2.3
